feat(row-counter): generate Python bindings
