+++ row-counter/app/pyproject.toml
@@ -1,0 +1,16 @@
+[tool.poetry]
+name = "row-counter"
+version = "0.0.0"
+description = ""
+authors = ["Nautilus"]
+license = "AGPL-3.0"
+
+[tool.poetry.dependencies]
+python = "^3.10"
+
+[tool.poetry.group.dev.dependencies]
+maturin = {extras = ["patchelf"], version = "^0.14.10"}
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
